requirements.txt
Django==5.2.5
langchain==0.3.27
langchain-community==0.3.27
transformers==4.55.2
torch==2.8.0
diffusers==0.34.0
accelerate==1.10.0
Pillow==11.3.0
opencv-python==4.12.0.88
librosa==0.11.0
soundfile==0.13.1
openai-whisper==20231117
python-dotenv==1.1.1
celery==5.5.3  # For background tasks
redis==6.2.0   # For caching